runtime-service/tests: unify Quarkus test profiles
This change moves a bunch of Quarkus test profiles into the shared `Profiles` class.

With that, exact duplicate profiles have been merged into one.

The Iceberg catalog profiles have been merged.

Each distinct Quarkus test profile requires the testing infrastructure to perform an expensive, time-consuming augmentation pass, hence this effort to unify test profiles.
