+++ runtime/service/src/test/java/org/apache/polaris/service/catalog/iceberg/LocalIcebergCatalogNoSqlInMemTest.java
@@ -18,29 +18,15 @@
  */
 package org.apache.polaris.service.catalog.iceberg;
 
-import static org.apache.polaris.service.catalog.Profiles.NOSQL_IN_MEM;
-
-import com.google.common.collect.ImmutableMap;
 import io.quarkus.test.junit.QuarkusTest;
 import io.quarkus.test.junit.TestProfile;
 import java.util.List;
-import java.util.Map;
 import org.apache.polaris.core.persistence.bootstrap.RootCredentialsSet;
+import org.apache.polaris.service.Profiles;
 
 @QuarkusTest
-@TestProfile(LocalIcebergCatalogNoSqlInMemTest.Profile.class)
+@TestProfile(Profiles.NoSqlIcebergCatalogProfile.class)
 public class LocalIcebergCatalogNoSqlInMemTest extends AbstractLocalIcebergCatalogTest {
-
-  public static class Profile extends AbstractLocalIcebergCatalogTest.Profile {
-    @Override
-    public Map<String, String> getConfigOverrides() {
-      return ImmutableMap.<String, String>builder()
-          .putAll(super.getConfigOverrides())
-          .putAll(NOSQL_IN_MEM)
-          .build();
-    }
-  }
-
   @Override
   protected void bootstrapRealm(String realmName) {
     metaStoreManagerFactory
