+++ runtime/service/src/test/java/org/apache/polaris/service/metrics/UserPrincipalTagEnabledMetricsTest.java
@@ -19,27 +19,9 @@
 package org.apache.polaris.service.metrics;
 
 import io.quarkus.test.junit.QuarkusTest;
-import io.quarkus.test.junit.QuarkusTestProfile;
 import io.quarkus.test.junit.TestProfile;
-import java.util.Map;
+import org.apache.polaris.service.Profiles;
 
 @QuarkusTest
-@TestProfile(UserPrincipalTagEnabledMetricsTest.Profile.class)
-public class UserPrincipalTagEnabledMetricsTest extends MetricsTestBase {
-
-  public static class Profile implements QuarkusTestProfile {
-
-    @Override
-    public Map<String, String> getConfigOverrides() {
-      return Map.of(
-          "polaris.metrics.tags.environment",
-          "prod",
-          "polaris.metrics.user-principal-tag.enable-in-api-metrics",
-          "true",
-          "polaris.metrics.realm-id-tag.enable-in-api-metrics",
-          "false",
-          "polaris.metrics.realm-id-tag.enable-in-http-metrics",
-          "false");
-    }
-  }
-}
+@TestProfile(Profiles.UserPrincipalTagEnabledMetricsProfile.class)
+public class UserPrincipalTagEnabledMetricsTest extends MetricsTestBase {}
